% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialReg.R
\name{BinomialReg}
\alias{BinomialReg}
\title{Ajuste de modelos de regressao multipla}
\usage{
BinomialReg(data,TwoWay=F,Family ="probit")
}
\arguments{
\item{data}{:Se houver apenas um fator (variavel explicativa) a matriz deve
 conter duas colunas. A primeira com os valores quantitativos da variavel
  explicativa e a segunda com os valores binarios (0 e 1) da variavel resposta. /cr
Se houver dois fatores, a matriz deve conter 3 colunas.Obrigatoriamente, as duas
primeiras devem ser com as variaveis explicativas e a terceira com a variavel resposta (dependente).}

\item{TwoWay}{:valor logico (TRUE ou FALSE). Coloque TRUE quando houver dois fatores
(variaveis explicativas) e FALSE se houver um unico fator.}

\item{Family}{:Identificacao do metodo utilizado: "probit" ou "logit".}
}
\value{
A funcao apresenta o resultado do ajuste de 16 modelos de regressao.
}
\description{
Esta funcao realiza o ajuste de 12 modelos de regressao multipla considerando 2 variaveis independentes (explicativas).
}
\examples{
###################################
############ Ajuste de uma regressao glm
data(Dados1)
D1=Dados1
BinomialReg(data=D1,TwoWay=FALSE,Family="probit")
}
\seealso{
/code{/link{aov}}, /code{/link{lm}}
}
