% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados1.R
\docType{data}
\name{Dados1}
\alias{Dados1}
\title{Dados de exemplo de um experimento com doses de um inseticida}
\format{
Um data frame dados de exemplo
}
\usage{
data(Dados1)
}
\description{
Exemplo simulado com dados de exemplo de um experimento com
 doses de um inseticida
}
\references{
<https://www.youtube.com/channel/UCDGyvLCJnv9RtTY1YMBMVNQ>
}
\author{
Alcinei Mistico Azevedo: \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
