% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DadosDAP.R
\docType{data}
\name{DadosDAP}
\alias{DadosDAP}
\title{Dados de exemplo de um experimento associando o DAP com a produtidade}
\format{
Um data frame dados de exemplo
}
\usage{
data(DadosDAP)
}
\description{
Dados de exemplo de um experimento associando o diametro a
altura do peito com a produtidade em Fava danta
}
\references{
<https://www.youtube.com/channel/UCDGyvLCJnv9RtTY1YMBMVNQ>
}
\author{
Alcinei Mistico Azevedo: \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
