% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DadosLeucena.R
\docType{data}
\name{DadosLeucena}
\alias{DadosLeucena}
\title{Dados de exemplo de um experimento com Leucaena Leucocephala}
\format{
Um data frame dados de exemplo
}
\usage{
data(DadosLeucena)
}
\description{
Dados de exemplo de um experimento com Leucaena Leucocephala
}
\references{
<https://www.youtube.com/channel/UCDGyvLCJnv9RtTY1YMBMVNQ>
}
\author{
Alcinei Mistico Azevedo: \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
