% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpGLM-package.R
\docType{package}
\name{ExpGLM-package}
\alias{ExpGLM-package}
\alias{ExpGLM}
\title{Generalized Linear Model for experiment analysis}
\description{
Package developed to facilitate the use of
generalized linear models in the analysis of experiments
(Pacote desenvolvido para facilitar  o uso de modelos
lineares generalizados na análise de experimentos).
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK4wSzwg-IxvrzaNzSLLrXEB>
}
\seealso{
\code{\link{BinomialReg}}
}
\author{
Alcinei Mistico Azevedo <alcineimistico@hotmail.com>
}
